feat: Save avatar and cover images

+++ dev/system/apis/account/core/src/graphql/mutation/account_profile_avatar.rs
@@ -13,7 +13,7 @@ use meower_shared::JwtClaims;
 
 use std::sync::Arc;
 
-use async_graphql::{ Context, Object, Result };
+use async_graphql::{ Context, Object, InputObject, Result };
 use base64::prelude::*;
 use object_store::ObjectStore;
 use object_store::path::Path as StoragePath;
@@ -26,6 +26,19 @@ use sea_orm::{
     ModelTrait,
     QueryFilter,
 };
+
+
+//------------------------------------------------------------------------------
+/// Inputs.
+//------------------------------------------------------------------------------
+#[derive(InputObject, Debug)]
+struct UploadAvatarInput
+{
+    account_profile_token: String,
+    file_name: Option<String>,
+    base64: Option<String>,
+    delete_file: bool,
+}
 
 
 //------------------------------------------------------------------------------
@@ -44,8 +57,7 @@ impl AccountProfileAvatarMutation
     (
         &self,
         ctx: &Context<'_>,
-        account_profile_token: String,
-        base64: Option<String>,
+        input: UploadAvatarInput,
     ) -> Result<bool>
     {
         let tsx = ctx.data::<Arc<DatabaseTransaction>>().unwrap().as_ref();
@@ -54,7 +66,7 @@ impl AccountProfileAvatarMutation
         let storage = ctx.data::<Arc<Box<dyn ObjectStore>>>().unwrap().as_ref();
 
         let account_profile = match AccountProfileEntity::find()
-            .filter(AccountProfileColumn::Token.eq(account_profile_token))
+            .filter(AccountProfileColumn::Token.eq(input.account_profile_token))
             .one(tsx)
             .await
             .unwrap()
@@ -77,21 +89,24 @@ impl AccountProfileAvatarMutation
             return Err(t!("system.error.unauthorized").into());
         }
 
-        if let Some(exists_avatar) = account_profile
-            .find_related(AccountProfileAvatarEntity)
-            .one(tsx)
-            .await
-            .unwrap()
+        if input.base64.is_some() || input.delete_file
         {
-            let exists_avatar_path = StoragePath::from
-            (
-                config.avatar_path.clone() + "/" + &exists_avatar.file_key
-            );
-            storage.delete(&exists_avatar_path).await.unwrap();
-            exists_avatar.delete(tsx).await.unwrap();
-        };
+            if let Some(exists_avatar) = account_profile
+                .find_related(AccountProfileAvatarEntity)
+                .one(tsx)
+                .await
+                .unwrap()
+            {
+                let exists_avatar_path = StoragePath::from
+                (
+                    config.avatar_path.clone() + "/" + &exists_avatar.file_key
+                );
+                storage.delete(&exists_avatar_path).await.unwrap();
+                exists_avatar.delete(tsx).await.unwrap();
+            };
+        }
 
-        if let Some(base64) = base64
+        if let Some(base64) = input.base64
         {
             let (prefix, base64) = match base64.split_once(",")
             {
@@ -113,6 +128,7 @@ impl AccountProfileAvatarMutation
             {
                 account_profile_id: ActiveValue::Set(account_profile_id),
                 content_type: ActiveValue::Set(content_type.to_string()),
+                file_name: ActiveValue::Set(input.file_name.unwrap_or_default()),
                 file_size: ActiveValue::Set(file_len),
                 ..Default::default()
             };
